// src/app/core/interfaces/idata.ts
export interface IData {
  name: string;
  position: string;
  profile: string;
  address: string;
  email: string;
  phone: string;
  techSkills:string[]
  skills: string[];
  expierence: string[];
  language: string[];
  education: string;
  educationDesc: string;
  image: string | ArrayBuffer | null;
}